A bug meant that "reboot" got saved as "current command" to eeprom, so the esp8266 would reboot, read the command to execute was "reboot" and... not be flashable.
Not sure I've fixed that bug, but a couple of safeguards now exist -
it ignores, clumsily, reboot upon reading from eeprom.  And if the
button is pressed at boot, it does not initialise things, so at least
it can still talk to the network long enough to be debugged, flashed.